+++ .gitattributes
@@ -1,2 +1,3 @@
 TestService.java.txt binary
+TestServiceLite.java.txt binary
 TestServiceNano.java.txt binary
